Move handleCreateTask to solveTask.ts
It makes more sense to have this function here than in the hooks.ts
since it is related to solving a task and not for hooking.
